+++ src/Aardvark.UI/Utilities/Common.fs
@@ -23,14 +23,18 @@ module MouseButtons =
         | 0 -> MouseButtons.Left
         | 1 -> MouseButtons.Middle
         | 2 -> MouseButtons.Right
+        | 3 -> MouseButtons.Button4
+        | 4 -> MouseButtons.Button5
         | _ -> MouseButtons.None
 
     /// See: https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/button
     let toEventButton (button: MouseButtons) =
         match button with
-        | MouseButtons.Left   -> 0
-        | MouseButtons.Middle -> 1
-        | MouseButtons.Right  -> 2
+        | MouseButtons.Left    -> 0
+        | MouseButtons.Middle  -> 1
+        | MouseButtons.Right   -> 2
+        | MouseButtons.Button4 -> 3
+        | MouseButtons.Button5 -> 4
         | _ -> 0
 
     /// See: https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/button
@@ -43,6 +47,8 @@ module MouseButtons =
         if buttons &&& 1 = 1 then &result |||= MouseButtons.Left
         if buttons &&& 2 = 2 then &result |||= MouseButtons.Right
         if buttons &&& 4 = 4 then &result |||= MouseButtons.Middle
+        if buttons &&& 8 = 8 then &result |||= MouseButtons.Button4
+        if buttons &&& 16 = 16 then &result |||= MouseButtons.Button5
         result
 
     /// See: https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/buttons
@@ -55,6 +61,8 @@ module MouseButtons =
         if buttons.HasFlag MouseButtons.Left then &result |||= 1
         if buttons.HasFlag MouseButtons.Right then &result |||= 2
         if buttons.HasFlag MouseButtons.Middle then &result |||= 4
+        if buttons.HasFlag MouseButtons.Button4 then &result |||= 8
+        if buttons.HasFlag MouseButtons.Button5 then &result |||= 16
         result
 
 module Server =